Integrate course correction with navigation logic in color_distance.py
Course correction is the same

Obstacle counter added. Can be incremented by pressing any key. Only certain colors will be looked for depending on counter (example, look for blues when counter = 2,4,6).

Course correction tells if vehicle is aligned. 
If not aligned, it also gives direction and degrees to turn. 
If aligned and close, tells vehicle to perform action
If aligned but far, just says vehicle is aligned
These will appear next to distance. Distance is in mm and "close" means 200 mm.