Create color_distance_ignore_yellow.py
Similar to color_distance. This one will ignore the yellow bucket (yellow bucket function is for ramp only). When it completes the last blue obstacle, it stops the vehicle. If it gets too close to an obstacle that is not aligned, will stop the vehicle. When vehicle is stopped, program will stop. This version doesn't stop when it detects no colors.